fix: make more modular, give application control of shown Pois
