+++ react-components/src/components/Architecture/pointsOfInterest/index.ts
@@ -2,6 +2,8 @@
  * Copyright 2024 Cognite AS
  */
 export { PoiInfoPanelContent } from './PoiInfoPanelContent';
-export { useSyncSelectedPoi } from './useSyncSelectedPoi';
+export { useFilterPointsOfInterest } from './useFilterPointsOfInterest';
 export { usePoiDomainObject } from './usePoiDomainObject';
+export { usePointsOfInterest } from './usePointsOfInterest';
+export { useSyncSelectedPoi } from './useSyncSelectedPoi';
 export { PoiList } from './PoiList';
